+++ client/bindings/nodejs/types/buildBlockOptions.ts
@@ -1,8 +1,10 @@
 // Copyright 2021-2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
-import type { IUTXOInput, OutputTypes } from '@iota/types';
+
+import type { IUTXOInput, OutputTypes, INativeToken } from '@iota/types';
 import type { CoinType } from '../lib';
 import type { IRange } from './range';
+import type { Burn } from './burn';
 
 /** Options to build a new block, possibly with payloads */
 export interface IBuildBlockOptions {
@@ -21,8 +23,8 @@ export interface IBuildBlockOptions {
     data?: string;
     /** Parent block IDs */
     parents?: string[];
-    /** Allow burning of native tokens */
-    allowBurning?: boolean;
+    /** Explicit burning of aliases, nfts, foundries and native tokens */
+    burn?: Burn;
 }
 
 /** Address with base coin amount */