+++ client/examples/output/foundry.rs
@@ -4,6 +4,7 @@
 //! cargo run --example foundry --release
 
 use iota_client::{
+    api::input_selection::Burn,
     block::{
         address::AliasAddress,
         output::{
@@ -24,7 +25,7 @@ use iota_client::{
 };
 use primitive_types::U256;
 
-/// In this example we will create an foundry output
+/// In this example we will create a foundry output
 
 #[tokio::main]
 async fn main() -> Result<()> {
@@ -85,6 +86,7 @@ async fn main() -> Result<()> {
     //////////////////////////////////////////////////
     // create foundry output and mint 70 native tokens
     //////////////////////////////////////////////////
+
     let alias_output_id = get_alias_output_id(block.payload().unwrap())?;
     let alias_id = AliasId::from(&alias_output_id);
     let token_scheme = TokenScheme::Simple(SimpleTokenScheme::new(
@@ -223,6 +225,7 @@ async fn main() -> Result<()> {
     //////////////////////////////////
     // send native token without foundry
     //////////////////////////////////
+
     let basic_output_id = get_basic_output_id_with_native_tokens(block.payload().unwrap())?;
     let outputs = vec![
         basic_output_builder
@@ -247,6 +250,7 @@ async fn main() -> Result<()> {
     //////////////////////////////////
     // burn native token without foundry
     //////////////////////////////////
+
     let basic_output_id = get_basic_output_id_with_native_tokens(block.payload().unwrap())?;
     let outputs = vec![
         basic_output_builder
@@ -257,7 +261,7 @@ async fn main() -> Result<()> {
     let block = client
         .block()
         .with_secret_manager(&secret_manager)
-        .with_burning_allowed(true)
+        .with_burn(Burn::new().add_native_token(token_id, 20))
         .with_input(basic_output_id.into())?
         .with_outputs(outputs)?
         .finish()